Comprehensive cell/tuft/den test

+++ src/axon_space.rs
@@ -91,7 +91,7 @@ mod tests {
 
 	pub trait AxonSpaceTest {
 		fn axn_state(&self, idx: usize) -> u8;
-		fn write_to_axon(&self, val: u8, idx: usize);
+		fn write_to_axon(&self, val: u8, idx: u32);
 	}
 
 	impl AxonSpaceTest for AxonSpace {
@@ -101,9 +101,9 @@ mod tests {
 			sdr[0]
 		}
 
-		fn write_to_axon(&self, val: u8, idx: usize) {
+		fn write_to_axon(&self, val: u8, idx: u32) {
 			let sdr = vec![val];
-			self.states.write_direct(&sdr, idx);
+			self.states.write_direct(&sdr, idx as usize);
 		}		
 	}
